added a default imgui layout to the editor layer

+++ Engine/src/Scene/Entity.h
@@ -37,7 +37,7 @@ public:
 		m_Scene->m_Registry.remove<T>(m_EntityHandle);
 	}
 
-	operator bool() const { return (uint32_t)m_EntityHandle != entt::null; }
+	operator bool() const { return m_EntityHandle != entt::null; }
 private:
 	entt::entity m_EntityHandle{ entt::null };
 	Scene* m_Scene = nullptr;
